Внесены правки, вместо целого числа дробные и убрала лишние строки

+++ triangle_area.rb
@@ -1,19 +1,15 @@
 # Площадь треугольника можно вычислить, зная его основание (a) и высоту (h) по формуле: 1/2*a*h.
 # Программа должна запрашивать основание и высоту треугольника и возвращать его площадь.
 
-# Запрашивать основание и высоту
-# Вычисляем площадь
-# Выводим площадь
-
 def ask_base
   puts "input the base: "
-  base = gets.to_i
+  base = gets.to_f
   return base
 end
 
 def ask_height
   puts "input the height: "
-  height = gets.to_i
+  height = gets.to_f
   return height
 end
 
@@ -21,8 +17,6 @@ def calculate_area(base,height)
   area = 0.5 * base * height
   return area
 end
-
-
 
 triangle_base = ask_base
 triangle_height = ask_height
